# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/transactions": {
      "get": {
        "operationId": "TransactionsController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      },
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/conflicts": {
      "get": {
        "operationId": "TransactionsController_conflicts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictReport"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_one",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      },
      "patch": {
        "operationId": "TransactionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}/confirm": {
      "post": {
        "operationId": "TransactionsController_confirm",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}/cancel": {
      "post": {
        "operationId": "TransactionsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}/pickup": {
      "post": {
        "operationId": "TransactionsController_pickup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{id}/return": {
      "post": {
        "operationId": "TransactionsController_returnItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
...
    "/api/transactions/{id}/checkout": {
      "post": {
        "operationId": "TransactionsController_checkout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Apollo Rigor API",
    "description": "Vendas e locações avulsas. Valores em centavos; datas YYYY-MM-DD. Checkout exclusivamente simulado. Auth intermediado pelo NestJS/Supabase.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "profileId": {
            "type": "string",
            "format": "uuid"
          },
          "variantId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "SALE",
              "RENTAL"
            ]
          },
          "priceCents": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          }
        },
        "required": [
          "profileId",
          "variantId",
          "type"
        ]
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "string",
            "format": "uuid"
          },
          "priceCents": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          }
        }
      },
      "ReturnDto": {
        "type": "object",
        "properties": {
          "damageNotes": {
            "type": "string",
            "maxLength": 2000
          }
        }
      },
...
      "Variant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "size": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "collection": {
            "type": "string"
          },
          "fabric": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "line": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          },
          "rentalPriceCents": {
            "type": "integer",
            "minimum": 0,
            "example": 10000
          },
          "salePriceCents": {
            "type": "integer",
            "minimum": 0,
            "example": 10000
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Variant"
            }
          }
        }
      },
      "Payment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "transactionId": {
            "type": "string",
            "format": "uuid"
          },
          "amountCents": {
            "type": "integer",
            "minimum": 0,
            "example": 10000
          },
          "simulated": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "CANCELLED",
              "REFUNDED"
            ]
          },
          "paidAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "profileId": {
            "type": "string",
            "format": "uuid"
          },
          "variantId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "SALE",
              "RENTAL"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "CONFIRMED",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "priceCents": {
            "type": "integer",
            "minimum": 0,
            "example": 10000
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-10-15"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-10-15"
          },
          "confirmedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "pickedUpAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "damageNotes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "payment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Payment"
              }
            ],
            "nullable": true
          },
          "variant": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Variant"
              }
            ],
            "properties": {
              "product": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        }
      },
...
      "ConflictReport": {
        "type": "object",
        "properties": {
          "overdue": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "conflicts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "transactionId": {
                  "type": "string",
                  "format": "uuid"
                },
                "overdueTransactionIds": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}